backend/requirements.txt
fastapi
uvicorn
pymongo
dnspython
python-dotenv
passlib
bcrypt
pyjwt
joblib
numpy
scikit-learn
pandas